requirements.txt
Flask==3.0.0
flask-cors==4.0.0
beautifulsoup4==4.12.2
requests==2.31.0
openai==1.3.0
python-dotenv==1.0.0
faiss-cpu==1.13.1
sentence-transformers>=2.2.2
numpy>=1.26.0
retrying==1.3.4
mcp>=1.0.0

